Updated code. Now uses Leaf objects to handle frequency and character.
Everything sorts using heap sort and it has already been done.. Currently doing the handling of the tree where.

The way the tree works is take the two first element  from the LeafArray, make a new node using the sum of the two element as its frequency and assign the two elements as left and right child.

Then, take the third item from the arrayLeaf and the node that has just been created and make a new node with the sum of their frequencies. and so on until you have one last node which becomes the root.

Then we have to do the compression by traversing the tree and assigning either 0 for the left branch and 1 with the right and output the codes to a file. Woot compression done.

Then we have to do decompression by putting the tree? in the header of the file and then like decompressing you know...

were so fucked woot

Also my old code has been commented out but its basically the same.

Also you are a fat fucking prick that totally loves cock and wishes he could get plates full of them.

Thank you very much

Agustin Suarez for Git News Laval

Have a good day